+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## [0.9.2] -  2020-09-24
 ### Added
 - add `programArgs` Set the inferior program arguments, to be used in the program run 
+- add `commandsBeforeExec` Commands run before execution.
 ## [0.9.1] -  2020-09-17
 ### Changed 
 -  Change keyword in `package.json`